requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.3
Flask-WTF==1.1.1
WTForms==3.0.1
Werkzeug==2.3.7
Jinja2==3.1.2
SQLAlchemy==2.0.23
python-dotenv==1.0.0
gunicorn==21.2.0
psycopg2-binary==2.9.7
email-validator==2.1.0
PyPDF2==3.0.1
python-multipart==0.0.6
Pillow==10.0.1